+++ config/control-ui-kit.php
@@ -91,9 +91,12 @@ return [
         'pill' => \ControlUIKit\Components\Pills\Pill::class,
 
         'table' => \ControlUIKit\Components\Tables\Table::class,
+        'table-action' => \ControlUIKit\Components\Tables\Action::class,
+        'table-action-options' => \ControlUIKit\Components\Tables\ActionOptions::class,
         'table-active-filter' => \ControlUIKit\Components\Tables\ActiveFilter::class,
         'table-cell' => \ControlUIKit\Components\Tables\Cell::class,
         'table-filter' => \ControlUIKit\Components\Tables\Filter::class,
+        'table-filters' => \ControlUIKit\Components\Tables\Filters::class,
         'table-row' => \ControlUIKit\Components\Tables\Row::class,
         'table-heading' => \ControlUIKit\Components\Tables\Heading::class,
         'table-empty' => \ControlUIKit\Components\Tables\EmptyRow::class,
@@ -240,6 +243,7 @@ return [
         'icon-search' => \ControlUIKit\Components\Icons\Search::class,
         'icon-search-cog' => \ControlUIKit\Components\Icons\SearchCog::class,
         'icon-shield-cross' => \ControlUIKit\Components\Icons\ShieldCross::class,
+        'icon-spinner' => \ControlUIKit\Components\Icons\Spinner::class,
         'icon-star-filled' => \ControlUIKit\Components\Icons\StarFilled::class,
         'icon-star' => \ControlUIKit\Components\Icons\Star::class,
         'icon-status-danger' => \ControlUIKit\Components\Icons\StatusDanger::class,
