+++ resources/js/control-ui-kit.js
@@ -153,100 +153,6 @@ window.Components = {
             onEscape() {
                 this.open = false
             },
-            onClickAway() {
-                this.open = false
-            },
-            onMoreButtonClicked() {
-                this.openMore = ! this.openMore
-                this.open = false
-            },
-            initFilters() {
-                let filter, width, ref, overflow, filters = this.$refs.filters;
-
-                if (! this.hasFilters) return;
-
-                for (let i = 0; i < filters.children.length; i++) {
-
-                    filter = filters.children[i]
-                    width = filter.offsetWidth
-
-                    if (filter.classList.contains('table-filter')) {
-
-                        ref = filter.dataset.ref
-                        overflow = document.querySelector('#more-filters [data-ref=' + ref + ']');
-                        overflow.classList.add('hidden')
-
-                        filter.id = 'filter_' + this.randomString()
-                        this.filters.push({
-                            id: filter.id,
-                            label: filter.dataset.label,
-                            priority: filter.dataset.priority,
-                            width: width,
-                            element: filter,
-                            overflow: overflow,
-                            location: 'filters'
-                        })
-
-                        filter.classList.add('hidden')
-                    }
-                }
-
-                this.filters.sort(this.sortByPriority);
-                this.resizeFilters(true)
-            },
-            randomString() {
-                return Date.now().toString(36) + Math.random().toString(36).substring(2)
-            },
-            resizeFilters(first) {
-                this.open = false
-                if (! this.hasFilters) return
-                let filter, used = 0
-                let searchFieldWidth = (this.hasSearch) ? this.$refs.search.offsetWidth : 0
-                let moreButtonWidth = this.$refs.more.offsetWidth
-                let tableWidth = this.$refs.table.offsetWidth
-                let maxWidth = tableWidth - searchFieldWidth - moreButtonWidth - 50
-
-                for (let i = 0; i < this.filters.length; i++) {
-
-                    filter = this.filters[i]
-                    if (first) filter.element.classList.remove('hidden')
-                    used += filter.width
-
-                    if ((used > maxWidth || window.innerWidth < 640) && filter.location !== 'more') {
-                        this.toggleFilter(filter, 'more');
-                        this.filters[i].location = 'more'
-                    } else if (used < maxWidth && filter.location !== 'filters') {
-                        this.toggleFilter(filter, 'filters');
-                        this.filters[i].location = 'filters'
-                    }
-                }
-
-                let moreFilterCount = this.filters.filter(function (filter) {
-                    return filter.location === 'more'
-                }).length
-
-                this.moreButton = moreFilterCount > 0
-                this.openMore = this.openMore ? this.moreButton : this.openMore
-            },
-            hasActiveFilters() {
-                if (this.$refs.active === undefined) {
-                    return false;
-                }
-
-                return this.$refs.active.children.length !== 0
-            },
-            tableWrapperClasses() {
-                return this.hasActiveFilters() ? this.withFilters : this.withoutFilters
-            },
-            toggleFilter(filter, to) {
-                if (to === 'more') {
-                    filter.element.classList.add('hidden')
-                    filter.overflow.classList.remove('hidden')
-                } else {
-                    filter.element.classList.remove('hidden')
-                    filter.overflow.classList.add('hidden')
-                }
-            },
             onKeyboardSelect() {
                 if (this.activeIndex !== null) this.highlightIndex = this.activeIndex
                 this.value = this.$refs['listbox-' + this.open].children[this.highlightIndex].dataset.value
