+++ resources/views/control-ui-kit/tables/action.blade.php
@@ -1,0 +1,3 @@
+<a {{ $attributes->merge($classes()) }}>
+    <x-dynamic-component :component="$icon" />
+</a>
