+++ resources/views/control-ui-kit/tables/active-filter.blade.php
@@ -1,6 +1,6 @@
-<div {{ $attributes->merge($classes()) }}>
+<div {{ $attributes->merge($classes())->only('class')  }}>
     <span>{{ $label }}</span>
-    <a href="{{ $href }}" class="flex items-center focus:outline-none focus:ring-0">
+    <a @if ($href) href="{{ $href }}" @endif class="flex items-center focus:outline-none focus:ring-0 cursor-pointer" {{ $attributes->except('class') }}>
         <x-dynamic-component :component="$icon" :styles="$iconStyles()" />
     </a>
 </div>
