+++ resources/views/control-ui-kit/tables/table.blade.php
@@ -1,104 +1,93 @@
-<div x-cloak
+<div
+{{--    x-cloak--}}
      x-data="Components.table({
-        hasFilters: {{ isset($filters) && $filters->isNotEmpty() ? 'true' : 'false' }},
-        hasSearch: {{ $showSearch() ? 'true' : 'false' }},
-        withFilters: '{{ $tableWrapperWithFilters() }}',
-        withoutFilters: '{{ $tableWrapperWithoutFilters() }}',
+{{--        hasFilters: {{ count($filters) ? 'true' : 'false' }},--}}
+{{--        hasSearch: {{ $showSearch() ? 'true' : 'false' }},--}}
+{{--        withFilters: '{{ $tableWrapperWithFilters() }}',--}}
+{{--        withoutFilters: '{{ $tableWrapperWithoutFilters() }}',--}}
         orderby: '{{ $orderby }}',
         sort: '{{ $sort }}',
       })"
-     x-on:resize.window="resizeFilters()"
-     x-on:resize.window="resizeFilters()"
-     @ready="initFilters()"
      x-init="init()"
-     @keydown.escape="onEscape()"
 >
-    @if ((isset($filters) && $filters->isNotEmpty()) || $showSearch())
+    @if ($showSearch())
 
-        <div @click.away="onClickAway()" @click.stop="onClickAway()">
+        <div>
 
             <div class="@if ($showSearch()){{ $searchBarSpacing }}@endif {{ $searchBar }}">
 
                 @if ($showSearch())
-                <div class="{{ $searchContainer }}" x-ref="search" @click="open = false">
-                    <x-form action="{{ $searchUrl() }}" method="get" name="{{ $searchFormName }}" id="{{ $searchFormName }}">
-                        <div class="{{ $searchWrapperClasses() }}">
-                            @if ($searchIcon())
-                            <div class="{{ $searchIconClasses() }}">
-                                <x-dynamic-component :component="$searchIcon()" :size="$searchIconSize()" />
-                            </div>
-                            @endif
-                            <input name="{{ $searchName }}"
-                                   type="{{ $searchType }}"
-                                   id="{{ $searchId }}"
-                                   value="{{ $search }}"
-                                   placeholder="{{ $searchPlaceholder }}"
-                                   @if ($wireSearch)
-                                   wire:model.debounce.500ms="search"
-                                   wire:keydown.prevent.enter=""
-                                   @else
-                                   onchange="document.search.submit();"
-                                   @endif
-                                   class="{{ $searchInputClasses() }}"
-                                   autocomplete="off"
-                                {{ $attributes->whereStartsWith('wire:') }}
-                            />
-                            @if ($wireSearch)
-                            <div wire:loading>
-                               loading
-                            </div>
-                            @endif
-                        </div>
-                    </x-form>
-                </div>
-                @endif
 
-                @if(isset($filters) && $filters->isNotEmpty())
-                    <div x-ref="container" class="{{ $tableFiltersContainer() }}">
-                        <div class="{{ $tableFiltersClasses() }}" x-ref="filters">
-                            {{ $filters }}
-                            <div>
-                                <button class="{{ $moreButtonClasses() }}" x-ref="more" x-show="moreButton" @click="onMoreButtonClicked()">
-                                    <x-dynamic-component :component="$moreButtonIcon()" :size="$moreButtonIconSize()" />
-                                </button>
+                    <div class="{{ $searchContainer }}" x-ref="search" @click="open = false">
+                        <x-form action="{{ $searchUrl() }}" method="get" name="{{ $searchFormName }}" id="{{ $searchFormName }}">
+                            <div class="{{ $searchWrapperClasses() }}">
+                                @if ($searchIcon())
+                                <div class="{{ $searchIconClasses() }}">
+                                    <x-dynamic-component :component="$searchIcon()" :size="$searchIconSize()" />
+                                </div>
+                                @endif
+                                <input name="{{ $searchName }}"
+                                       type="{{ $searchType }}"
+                                       id="{{ $searchId }}"
+                                       value="{{ $search }}"
+                                       placeholder="{{ $searchPlaceholder }}"
+                                       @if ($wireSearch)
+                                       wire:model.debounce.500ms="search"
+                                       wire:keydown.prevent.enter=""
+                                       @else
+                                       onchange="document.search.submit();"
+                                       @endif
+                                       class="{{ $searchInputClasses() }}"
+                                       autocomplete="off"
+                                    {{ $attributes->whereStartsWith('wire:') }}
+                                />
+                                @if ($wireSearch)
+                                <div wire:loading>
+                                   <x-dynamic-component component="icon-spinner" class="mr-2 text-orange-500" />
+                                </div>
+                                @endif
                             </div>
-                        </div>
+                        </x-form>
+
                     </div>
-                @else
-                    <div class="{{ $tableFiltersEmpty() }}"></div>
+
+                    @if(count($filters))
+
+                        <x-table-filters :filters="$filters" />
+
+                    @endif
+
+{{--                <button class="{{ $moreButtonClasses() }}">--}}
+{{--                    <x-dynamic-component component="icon-download" :size="$moreButtonIconSize()" color="text-gray-700" />--}}
+{{--                </button>--}}
                 @endif
 
             </div>
-
-            @if(isset($filters) && $filters->isNotEmpty())
-                <div class="{{ $moreFiltersWrapper() }}">
-                    <div id="more-filters" x-ref="overflow" x-show="openMore" class="{{ $moreFiltersClasses() }}">
-                        {{ $filters }}
-                    </div>
-                </div>
-            @endif
 
         </div>
 
-        @if(isset($active) && $active->isNotEmpty())
+        @if($hasActiveFilters())
             <div class="{{ $activeFilterWrapperClasses() }}">
-                <div class="{{ $activeFilterListClasses() }}" x-ref="active">
-                    {{ $active }}
+                <div class="{{ $activeFilterListClasses() }}">
+                    @foreach ($activeFilters() as $filter)
+{{--                        wire:click="clearFilter('{{ $filter['name'] }}')"--}}
+                        <x-table-active-filter :name="$filter['name']" :label="$filter['label']"  x-on:click="$dispatch('clear-single-filter', '{{ $filter['name'] }}' )" />
+                    @endforeach
                 </div>
 
-                <a class="{{ $clearFilterClasses() }}"
-                   @if($clearFiltersHref) href="{{ $clearFiltersHref }}" @endif
-                @if($clearFiltersEvent) {!! $clearFiltersEvent !!} @endif
+                <a class="{{ $clearFilterClasses() }} cursor-pointer"
+                   x-on:click="$dispatch('clear-filters')"
+
+{{--                   @if($clearFiltersHref) href="{{ $clearFiltersHref }}" @endif--}}
+{{--                    @if($clearFiltersEvent) {!! $clearFiltersEvent !!} @endif--}}
                 >{{ $clearFiltersText }}</a>
 
             </div>
-        @else
-            <div x-ref="active"></div>
         @endisset
 
     @endif
 
-    <div class="{{ $tableWrapperClasses }}" :class="tableWrapperClasses()" x-ref="table">
+    <div class="{{ $tableWrapperClasses }}">
         <table {{ $attributes->merge($classes())->whereDoesntStartWith('wire:model') }}>
             @isset($headings)
             <thead>
