+++ src/Components/Icons/Spinner.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace ControlUIKit\Components\Icons;
+
+class Spinner extends BaseIconComponent
+{
+    public function render()
+    {
+        return view('control-ui-kit::control-ui-kit.icons.spinner');
+    }
+}
